Fixes Engineer Voidsuit
The old paths pointed to non-existent icons other than the technosuit-void, causing the ground sprites to be invisible, or for the onmob and ground sprites to be invisible. The fix is a stopgap mostly just using the existing technosuit and technohelmet.

In other words, either revert this to the old engineering suit already in the code, or make the new sprites.